added pages and components, and organized them in the right folder
